refactor(test-cleanup): hoist test-cache clearers into leaf modules
`cleanupTestDatabase` used three `await import()` calls to clear
process-wide caches (MCP session map, in-memory MCP transport sessions,
multi-tenant auth TTL caches). Each dynamically-loaded module
transitively pulled in the entire server app graph — most importantly
`@lobu/connector-sdk`, a workspace package whose dist is only built by
`make build-packages`.

After WS9 wired more revocation call sites into the gateway, the
`Security tests (bun:test)` job (which builds only `packages/core`)
started failing in `revoked-token-store.test.ts` with:

  Cannot find module '@lobu/connector-sdk' from
    .../packages/server/src/identity/capability-registry.ts
  Cannot find module '@lobu/connector-sdk' from
    .../packages/server/src/tools/save_content.ts

The user explicitly reverted adding `make build-packages` to the
security workflow; this is the slim alternative.

Each clearer is now exported from a dedicated leaf module that imports
only what it owns (a Map and a `clear()` call):

- `packages/server/src/__tests__/setup/mcp-session-cache.ts` — test-only
  MCP session map (previously inside `test-helpers.ts`, which loads
  `../../index`).
- `packages/server/src/mcp-session-state.ts` — the in-memory MCP
  transport session Map (previously inside `mcp-handler.ts`, which loads
  the entire tool registry).
- `packages/server/src/workspace/multi-tenant-caches.ts` — the four
  auth TTL caches (previously inside `multi-tenant.ts`, which loads the
  full auth stack).

`mcp-handler.ts` and `multi-tenant.ts` now read from the same shared
Map / cache instances and re-export the test clearer so existing call
sites keep working. `cleanupTestDatabase` switches to static imports
of the three leaf modules — no dynamic import, no policy violation.

Verification (matches what the Security tests workflow runs):

  cd packages/core && bun run build && cd ../..
  bun test packages/server/src/gateway/__tests__/revoked-token-store.test.ts
  bun test packages/server/src/gateway/__tests__/proxy-hardening.test.ts

Both pass without building any other workspace package. Full
`packages/server/src/gateway/__tests__/` (789 tests) and
`packages/server/src/__tests__/unit/` (166 tests) also pass.

Logical paths still 401 a revoked jti — no enforcement change, only
the test-cleanup import graph shrunk.
